requirements.txt
discord.py
PyNaCl==1.5.0
pandas
dnspython==1.16.0
async-timeout==3.0.1
